requirements.txt
-f /usr/share/pip-wheels
asgiref==3.2.10
beautifulsoup4==4.8.2
certifi==2020.6.20
chardet==3.0.4
cssselect==1.1.0
Django==3.0.9
django-debug-toolbar==2.2
django-extensions==3.0.5
django-filter==2.3.0
django-mathfilters==1.0.0
django-modelcluster==5.0.2
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.11.1
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
html5lib==1.1
idna==2.10
jdcal==1.4.1
l18n==2018.5
lxml==4.5.2
markdown2==2.3.9
mysqlclient==2.0.1
openpyxl==3.0.4
packaging==20.4
Pillow==7.2.0
pyparsing==2.4.7
pyquery==1.4.1
python-dotenv==0.14.0
pytz==2020.1
readtime==1.1.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tablib==2.0.0
Unidecode==1.1.1
urllib3==1.25.10
wagtail==2.9.3
webencodings==0.5.1
whitenoise==5.2.0
Willow==1.3
xlrd==1.2.0
XlsxWriter==1.3.3
xlwt==1.3.0
